third_party/WebKit/LayoutTests/http/tests/origin_trials/webexposed/webusb-origin-trial-interfaces-script-added.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>WebUSB - origin trial is enabled by script-added meta tag</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../resources/origin-trials-helper.js"></script>
<script>
// Generate token with the command:
// generate_token.py  http://127.0.0.1:8000 WebUSB2 --expire-timestamp=2000000000
let token = "AmgWWAL+v07ElzPwp9oXvY1ldvKTbMlEUFopboMutESil7WUfQNVdYFUzfai6GRq5MndTa+ln+w1gLlRw8VshwAAAABPeyJvcmlnaW4iOiAiaHR0cDovLzEyNy4wLjAuMTo4MDAwIiwgImZlYXR1cmUiOiAiV2ViVVNCMiIsICJleHBpcnkiOiAyMDAwMDAwMDAwfQ==";

let properties_to_check = {'Navigator': ['usb']};
let interfaces_to_check = ['USBAlternateInterface',
    'USBConfiguration',
    'USBConnectionEvent',
    'USBDevice',
    'USBEndpoint',
    'USBInTransferResult',
    'USBInterface',
    'USBIsochronousInTransferPacket',
    'USBIsochronousInTransferResult',
    'USBIsochronousOutTransferPacket',
    'USBIsochronousOutTransferResult',
    'USBOutTransferResult'];

test(t => {
  OriginTrialsHelper.check_properties_missing(this, properties_to_check);
}, "WebUSB related properties on interfaces before adding trial token via script.");

test(t => {
  OriginTrialsHelper.check_interfaces_missing(this, interfaces_to_check);
}, "WebUSB related interfaces before adding trial token via script.");

OriginTrialsHelper.add_token(token);

test(t => {
  OriginTrialsHelper.check_properties(this, properties_to_check);
}, "WebUSB related properties on interfaces after adding trial token via script.");

test(t => {
  OriginTrialsHelper.check_interfaces(this, interfaces_to_check);
}, "WebUSB related interfaces after adding trial token via script.");
</script>
