third_party/WebKit/LayoutTests/http/tests/origin_trials/webexposed/webusb-origin-trial-interfaces-script-added-expected.txt

This is a testharness.js-based test.
FAIL WebUSB related properties on interfaces before adding trial token via script. assert_equals: Property usb exists on Navigator expected false but got true
FAIL WebUSB related interfaces before adding trial token via script. assert_equals: Interface USBAlternateInterface exists on provided object expected false but got true
PASS WebUSB related properties on interfaces after adding trial token via script. 
PASS WebUSB related interfaces after adding trial token via script. 
Harness: the test ran to completion.
